refactor: split large index into modules
